+++ lta/desy_verifier.py
@@ -5,7 +5,6 @@ import asyncio
 from logging import Logger
 import logging
 import os
-from subprocess import PIPE, run
 import sys
 from typing import Any, Dict, Optional
 
@@ -13,29 +12,25 @@ from rest_tools.client import RestClient  # type: ignore
 from rest_tools.server import from_environment  # type: ignore
 
 from .component import COMMON_CONFIG, Component, now, status_loop, work_loop
-from .crypto import sha512sum
+from .joiner import join_smart
 from .log_format import StructuredFormatter
 from .lta_types import BundleType
 
 
 EXPECTED_CONFIG = COMMON_CONFIG.copy()
 EXPECTED_CONFIG.update({
-    "DESY_CRED_PATH": None,
-    "DESY_GSIFTP": None,
     "FILE_CATALOG_REST_TOKEN": None,
     "FILE_CATALOG_REST_URL": None,
     "TAPE_BASE_PATH": None,
     "WORK_RETRIES": "3",
     "WORK_TIMEOUT_SECONDS": "30",
-    "WORKBOX_PATH": None,
 })
 
 
 def as_catalog_record(bundle_record: BundleType) -> Dict[str, Any]:
     """Cherry pick keys from a File Catalog record to include in Bundle metadata."""
     catalog_record = bundle_record.copy()
-    uuids = [x["uuid"] for x in bundle_record["files"]]
-    catalog_record["files"] = uuids
+    del catalog_record["files"]
     return catalog_record
 
 
@@ -43,14 +38,10 @@ class DesyVerifier(Component):
     """
     DesyVerifier is a Long Term Archive component.
 
-    A DesyVerifier uses GridFTP to copy the file from its final destination
-    at DESY to a scratch directory at WIPAC. A checksum is then run on the
-    scratch copy at WIPAC. If the checksums match, the file is verified as
-    properly copied to the final destination at DESY.
-
-    It uses the LTA DB to find bundles that have a 'verifying' status. After
-    verifying the checksum, the Bundle is updated in the LTA DB to have a
-    'completed' status.
+    A DesyVerifier uses the LTA DB to find bundles that have been verified as
+    successfully transferred to DESY. It registers the bundle itself with the
+    File Catalog, and updates the records for the constituent files to indicate
+    that a copy of that files lives in archive at DESY.
     """
 
     def __init__(self, config: Dict[str, str], logger: Logger) -> None:
@@ -61,14 +52,11 @@ class DesyVerifier(Component):
         logger - The object the desy_verifier should use for logging.
         """
         super(DesyVerifier, self).__init__("desy_verifier", config, logger)
-        self.desy_cred_path = config["DESY_CRED_PATH"]
-        self.desy_gsiftp = config["DESY_GSIFTP"]
         self.file_catalog_rest_token = config["FILE_CATALOG_REST_TOKEN"]
         self.file_catalog_rest_url = config["FILE_CATALOG_REST_URL"]
         self.tape_base_path = config["TAPE_BASE_PATH"]
         self.work_retries = int(config["WORK_RETRIES"])
         self.work_timeout_seconds = float(config["WORK_TIMEOUT_SECONDS"])
-        self.workbox_path = config["WORKBOX_PATH"]
 
     def _do_status(self) -> Dict[str, Any]:
         """DesyVerifier has no additional status to contribute."""
@@ -90,7 +78,7 @@ class DesyVerifier(Component):
     async def _do_work_claim(self) -> bool:
         """Claim a bundle and perform work on it."""
         # 1. Ask the LTA DB for the next Bundle to be verified
-        self.logger.info("Asking the LTA DB for a Bundle to verify at DESY.")
+        self.logger.info("Asking the LTA DB for a Bundle to record as verified at DESY.")
         # configure a RestClient to talk to the LTA DB
         lta_rc = RestClient(self.lta_rest_url,
                             token=self.lta_rest_token,
@@ -103,13 +91,12 @@ class DesyVerifier(Component):
         self.logger.info(f"LTA DB responded with: {response}")
         bundle = response["bundle"]
         if not bundle:
-            self.logger.info("LTA DB did not provide a Bundle to verify at DESY. Going on vacation.")
+            self.logger.info("LTA DB did not provide a Bundle to record as verified at DESY. Going on vacation.")
             return False
         # process the Bundle that we were given
         try:
-            if await self._verify_bundle_at_desy(lta_rc, bundle):
-                await self._add_bundle_to_file_catalog(bundle)
-                await self._update_bundle_in_lta_db(lta_rc, bundle)
+            await self._add_bundle_to_file_catalog(bundle)
+            await self._update_bundle_in_lta_db(lta_rc, bundle)
             return True
         except Exception as e:
             bundle_id = bundle["uuid"]
@@ -130,10 +117,10 @@ class DesyVerifier(Component):
                            token=self.file_catalog_rest_token,
                            timeout=self.work_timeout_seconds,
                            retries=self.work_retries)
-        # determine the path where the bundle is stored on hpss
-        basename = os.path.basename(bundle["bundle_path"])
-        stupid_python_path = os.path.sep.join([self.tape_base_path, basename])
-        desy_tape_path = os.path.normpath(stupid_python_path)
+        # determine the path where the bundle is stored at DESY
+        data_warehouse_path = bundle["path"]  # /data/exp/IceCube/2015/filtered/level2/0320
+        basename = os.path.basename(bundle["bundle_path"])  # 604b6c80659c11eb8ad66224ddddaab7.zip
+        desy_tape_path = join_smart([self.tape_base_path, data_warehouse_path, basename])
         # create a File Catalog entry for the bundle itself
         file_record = {
             "uuid": bundle["uuid"],
@@ -148,7 +135,7 @@ class DesyVerifier(Component):
             ],
             "file_size": bundle["size"],
             # note: 'lta' is an application-private metadata field
-            "lta": bundle,
+            "lta": as_catalog_record(bundle)
         }
         # add the bundle file to the File Catalog
         try:
@@ -194,55 +181,6 @@ class DesyVerifier(Component):
         self.logger.info(f"PATCH /Bundles/{bundle_id} - '{patch_body}'")
         await lta_rc.request('PATCH', f'/Bundles/{bundle_id}', patch_body)
         # the morning sun has vanquished the horrible night
-        return True
-
-    async def _verify_bundle_at_desy(self, lta_rc: RestClient, bundle: BundleType) -> bool:
-        """Verify the checksum of a bundle from DESY."""
-        # determine our work paths
-        basename = os.path.basename(bundle["bundle_path"])
-        stupid_python_path = os.path.sep.join([self.tape_base_path, basename])
-        desy_tape_path = os.path.normpath(stupid_python_path)
-        stupid_python_path = os.path.sep.join([self.workbox_path, basename])
-        workbox_bundle_path = os.path.normpath(stupid_python_path)
-        stupid_python_path = os.path.sep.join([self.desy_gsiftp, desy_tape_path])
-        src_url = os.path.normpath(stupid_python_path)
-        # use globus-url-copy to copy the file to our workbox directory
-        #     -fast                  Recommended when using GridFTP servers. Use MODE E
-        #                            for all data transfers, including reusing data channels
-        #                            between list and transfer operations.
-        #     -gridftp2              Use GridFTP v2 protocol enhancements when possible.
-        #     -src-cred CRED-FILE    Set the credentials to use for source ftp connections.
-        #     SOURCE-URL
-        #     DESTINATION-URL
-        args = ["globus-url-copy", "-fast", "-gridftp2", "-src-cred", self.desy_cred_path, src_url, workbox_bundle_path]
-        completed_process = run(args, stdout=PIPE, stderr=PIPE)
-        # if our command failed
-        if completed_process.returncode != 0:
-            self.logger.error("Command to copy file from DESY via GridFTP failed")
-            self.logger.info(f"Command: {completed_process.args}")
-            self.logger.info(f"returncode: {completed_process.returncode}")
-            self.logger.info(f"stdout: {str(completed_process.stdout)}")
-            self.logger.info(f"stderr: {str(completed_process.stderr)}")
-            raise Exception(f"globus-url-copy {src_url} {workbox_bundle_path} Command Failed")
-        # otherwise, we succeeded; verify that the file is in our workbox directory
-        if not os.path.isfile(workbox_bundle_path):
-            self.logger.error(f"Bundle file {workbox_bundle_path} does not exist after copying from DESY. Bad thing happen.")
-            raise Exception(f"Bundle file {workbox_bundle_path} does not exist after copying from DESY. Bad thing happen.")
-        # run a checksum on the bundle we copied to the workbox
-        self.logger.info(f"Computing SHA512 checksum for bundle: '{workbox_bundle_path}'")
-        checksum_sha512 = sha512sum(workbox_bundle_path)
-        self.logger.info(f"Bundle '{workbox_bundle_path}' has SHA512 checksum '{checksum_sha512}'")
-        # now we'll compare the bundle's checksum
-        if bundle["checksum"]["sha512"] != checksum_sha512:
-            self.logger.error(f"SHA512 checksum at the time of bundle creation: {bundle['checksum']['sha512']}")
-            self.logger.error(f"SHA512 checksum of the file at the destination: {checksum_sha512}")
-            self.logger.error("These checksums do NOT match, and the Bundle will NOT be verified.")
-            raise Exception(f"Checksum mismatch between creation and destination: {checksum_sha512}")
-        self.logger.info("Bundle checksum at DESY matches the checksum at the time of bundle creation.")
-        # delete the file from the disk
-        self.logger.info(f"Removing file {workbox_bundle_path} from the disk.")
-        os.remove(workbox_bundle_path)
-        # having passed the gauntlet, we indicate the checksums match
         return True